Move Input related stuff from problem_startsolutions into Input.input
