+++ src/input.jl
@@ -2,6 +2,7 @@ module Input
 
 import MultivariatePolynomials
 const MP = MultivariatePolynomials
+using LinearAlgebra
 
 import ..Homotopies
 import ..Systems
@@ -24,24 +25,26 @@ const Inputs = Union{Systems.AbstractSystem, MPPolys}
 
 Construct a `StartTargetProblem` out of two systems of polynomials.
 """
-struct StartTarget{P1<:Inputs, P2<:Inputs, V<:AbstractVector} <: AbstractInput
+struct StartTarget{P1<:Inputs, P2<:Inputs} <: AbstractInput
     start::P1
     target::P2
-    startsolutions::Vector{V}
+    startsolutions
 
-    function StartTarget{P1, P2, V}(start::P1, target::P2, startsolutions::Vector{V}) where {P1<:Inputs, P2<:Inputs, V<:AbstractVector}
-        @assert length(start) == length(target) "Cannot construct `StartTargetProblem` since the lengths of `start` and `target` don't match."
+    function StartTarget{P1, P2}(start::P1, target::P2, startsolutions) where {P1<:Inputs, P2<:Inputs}
+        if length(start) ≠ length(target)
+            error("Cannot construct `StartTargetProblem` since the lengths of `start` and `target` don't match.")
+        end
         new(start, target, startsolutions)
     end
 end
-function StartTarget(start::P1, target::P2, startsolutions::Vector{V}) where {P1<:Inputs, P2<:Inputs, V<:AbstractVector}
-    StartTarget{P1, P2, V}(start, target, startsolutions)
+function StartTarget(start::P1, target::P2, startsolutions) where {P1<:Inputs, P2<:Inputs}
+    StartTarget{P1, P2}(start, target, startsolutions)
 end
 
 
-struct Homotopy{Hom<:Homotopies.AbstractHomotopy, V} <: AbstractInput
+struct Homotopy{Hom<:Homotopies.AbstractHomotopy} <: AbstractInput
     H::Hom
-    startsolutions::Vector{V}
+    startsolutions
 end
 
 
@@ -53,6 +56,10 @@ is the target system and a total degree system should be assembled.
 """
 struct TotalDegree{S<:Inputs} <: AbstractInput
     system::S
+    degrees::Vector{Int}
+end
+function TotalDegree(S::Vector{<:MP.AbstractPolynomialLike})
+    TotalDegree(S, MP.maxdegree.(S))
 end
 
 
@@ -61,13 +68,130 @@ end
 
 Construct a `ParameterSystem`. This indicates that the system `system` has variables `x` and parameters `p`.
 """
-struct ParameterSystem{P<:MP.AbstractPolynomialLike, V<:MP.AbstractVariable, T1<:Number, T2<:Number, V1<:AbstractVector} <: AbstractInput
+struct ParameterSystem{P<:MP.AbstractPolynomialLike, V<:MP.AbstractVariable} <: AbstractInput
     system::Vector{P}
     parameters::Vector{V}
-    p₁::Vector{T1}
-    p₀::Vector{T2}
-    startsolutions::Vector{V1}
+    p₁::Vector
+    p₀::Vector
+    startsolutions
+    y₁::ComplexF64
+    y₀::ComplexF64
 end
 
+const overdetermined_error_msg = """
+The input system is overdetermined. Therefore it is necessary to provide an explicit start system.
+See
+    https://www.JuliaHomotopyContinuation.org/guides/latest/overdetermined_tracking/
+for details.
+"""
+
+
+const supported_keywords = [:parameters, :start_parameters, :target_parameters,
+                            :target_gamma, :start_gamma, :p₁, :p₀, :γ₁, :γ₀]
+
+"""
+    input(F::Vector{<:MP.AbstractPolynomial})::TotalDegree
+    input(F::Systems.AbstractSystem)::TotalDegree
+    input(G::Vector{<:MP.AbstractPolynomial}, F::Vector{<:MP.AbstractPolynomial}, startsolutions)::StartTargetProblem
+    input(F::Vector{<:MP.AbstractPolynomial}, parameters, startsolutions; kwargs...)::ParameterSystem
+    input(H::Homotopies.AbstractHomotopy, startsolutions)::Homotopy
+
+Construct an `AbstractInput`.
+"""
+function input(F::Vector{<:MP.AbstractPolynomial})
+    F = filter(f -> !iszero(f), F)
+    check_zero_dimensional(F)
+    # square system and each polynomial is non-zero
+    if length(F) == MP.nvariables(F) && ishomogenous(F)
+        error("The input system is a square homogenous system. This will result in an at least 1 dimensional solution space.")
+    end
+    TotalDegree(F, MP.maxdegree.(F))
+end
+
+function input(F::Systems.AbstractSystem)
+    n, N = size(F)
+    degrees = check_homogenous_degrees(F)
+    # system needs to be homogenous
+    if n + 1 > N
+        error(overdetermined_error_msg)
+    elseif  n + 1 ≠ N
+        error("Input system is not a square homogenous system!")
+    end
+    TotalDegree(F, degrees)
+end
+
+function input(G::Vector{<:MP.AbstractPolynomial}, F::Vector{<:MP.AbstractPolynomial}, startsolutions)
+    if length(G) ≠ length(F)
+        error("Start and target system don't have the same length")
+    end
+    check_zero_dimensional(F)
+    StartTarget(G, F, startsolutions)
+end
+
+function input(F::Vector{<:MP.AbstractPolynomial}, startsolutions;
+    parameters::Vector{<:MP.AbstractVariable}=error("parameters notdefined"),
+    start_parameters=randn(ComplexF64, length(parameters)), p₁ = start_parameters,
+    target_parameters=randn(ComplexF64, length(parameters)), p₀ = target_parameters,
+    start_gamma=randn(ComplexF64), γ₁ = start_gamma,
+    target_gamma=randn(ComplexF64), γ₀ = target_gamma)
+
+    if !(length(parameters) == length(p₁) == length(p₀))
+        error("Number of parameters doesn't match!")
+    end
+
+    ParameterSystem(F, parameters, p₁, p₀, startsolutions, γ₁, γ₀)
+end
+
+function input(H::Homotopies.AbstractHomotopy, startsolutions)
+    check_homogenous_degrees(Systems.FixedHomotopy(H, rand()))
+    Homotopy(H, startsolutions)
+end
+
+
+"""
+    check_homogenous_degrees(F::AbstractSystem)
+
+Compute (numerically) the degrees of `F` and verify that `F` is homogenous,
+"""
+function check_homogenous_degrees(F::Systems.AbstractSystem)
+    n, N = size(F)
+    if n < N - 1
+        error("Input system is not homogenous! It has $n polynomials in $N variables according to `size`.")
+    end
+    # The number of variables match, but it still cannot be homogenous.
+    # We evaluate the system with y:=rand(N) and 2y. If homogenous then the output
+    # scales accordingly to the degrees which we can obtain by taking logarithms.
+    x = rand(ComplexF64, N)
+    cache = Systems.cache(F, x)
+    y = Systems.evaluate(F, x, cache)
+    rmul!(x, 2)
+    y2 = Systems.evaluate(F, x, cache)
+
+    degrees = map(y2, y) do y2ᵢ, yᵢ
+        # y2ᵢ = 2^dᵢ yᵢ
+        float_dᵢ = log2(abs(y2ᵢ / yᵢ))
+        dᵢ = round(Int, float_dᵢ)
+        if abs(dᵢ - float_dᵢ) > 1e-10
+            error("Input system is not homogenous by our numerical check.")
+        end
+        dᵢ
+    end
+    degrees
+end
+
+"""
+    check_zero_dimensional(F::Vector{<:MP.AbstractPolynomial})
+
+Check that the given polynomial system can have zero dimensional components.
+"""
+function check_zero_dimensional(F::Vector{<:MP.AbstractPolynomial})
+    N = MP.nvariables(F)
+    n = length(F)
+
+    if n ≥ N || (n == N - 1 && ishomogenous(F))
+        return nothing
+    end
+    error("The input system will not result in a finite number of solutions.")
+end
 
 end
