+++ src/problems/totaldegree_solutions.jl
@@ -1,0 +1,66 @@
+"""
+    totaldegree_solutions(F::Vector{<:MP.AbstractPolynomialLike})
+
+Returns an iterator of the solutions of the total degree startsystem of `F`.
+"""
+function totaldegree_solutions(F::Vector{AP}, homogenization) where {AP<:MP.AbstractPolynomialLike}
+    totaldegree_solutions(MP.maxdegree.(F), homogenization)
+end
+function totaldegree_solutions(degrees::Vector{Int}, ::NullHomogenization)
+    TotalDegreeSolutionIterator(degrees, true)
+end
+function totaldegree_solutions(degrees::Vector{Int}, ::Homogenization)
+    TotalDegreeSolutionIterator(degrees, false)
+end
+
+
+"""
+    TotalDegreeSolutionIterator(degrees, homogenous::Bool)
+
+Given the `Vector`s `degrees` `TotalDegreeSolutionIterator` enumerates all solutions
+of the system
+```math
+\\begin{align*}
+    z_1^{d_1} - 1 &= 0 \\\\
+    z_1^{d_2} - 1 &= 0 \\\\
+    &\\vdots \\\\
+    z_n^{d_n} - 1 &= 0 \\\\
+\\end{align*}
+```
+where ``d_i`` is `degrees[i]`. If `homogenous` is `true` then
+the solutions of the system will be embedded in projective space by the map
+`x → [1.0; x]`.
+"""
+struct TotalDegreeSolutionIterator{Iter}
+    degrees::Vector{Int}
+    homogenous::Bool
+    iterator::Iter
+end
+function TotalDegreeSolutionIterator(degrees::Vector{Int}, homogenous::Bool)
+    iterator = Base.Iterators.product(map(d -> 0:d-1, degrees)...)
+    TotalDegreeSolutionIterator(degrees, homogenous, iterator)
+end
+
+function Base.iterate(iter::TotalDegreeSolutionIterator)
+    indices, state = iterate(iter.iterator)
+    _value(iter, indices), state
+end
+function Base.iterate(iter::TotalDegreeSolutionIterator, state)
+    it = iterate(iter.iterator, state)
+    it === nothing && return nothing
+    _value(iter, it[1]), it[2]
+end
+
+function _value(iter::TotalDegreeSolutionIterator, indices)
+    value = Vector{Complex{Float64}}()
+    if iter.homogenous
+        push!(value, complex(1.0, 0.0))
+    end
+    for (i, k) in enumerate(indices)
+        push!(value, cis(2π * k / iter.degrees[i]))
+    end
+    value
+end
+
+Base.length(iter::TotalDegreeSolutionIterator) = length(iter.iterator)
+Base.eltype(iter::Type{<:TotalDegreeSolutionIterator}) = Vector{Complex{Float64}}
