+++ test/problem_test.jl
@@ -17,31 +17,33 @@
 
     @polyvar x y z
     F = [x^2+y^2+2z^2, x+y+3z]
-    P, start = Problems.problem_startsolutions(Input.TotalDegree(F), homvar=y)
+    P, _ = Problems.problem_startsolutions(Input.TotalDegree(F), homvar=y)
     @test Problems.homogenization(P) == Problems.Homogenization(2)
 
-    P, start = Problems.problem_startsolutions(Input.TotalDegree(F))
+    P, _ = Problems.problem_startsolutions(Input.TotalDegree(F))
     @test Problems.homogenization(P) == Problems.NullHomogenization()
 
     @test_throws ErrorException Problems.problem_startsolutions(Input.StartTarget(
         [x^2+y^2+z^2, x^4+y^4+z^3], [x^3+z^3,y^3-z^3], [rand(ComplexF64, 3)]))
 
-    P, start = Problems.problem_startsolutions(Input.StartTarget(
+    P, _ = Problems.problem_startsolutions(Input.StartTarget(
         [x^2+y^2+z^2, x^4+y^4+z^4], [x^3+z^3,y^3-z^3], [rand(ComplexF64, 3)]))
     @test Problems.homogenization(P) == Problems.NullHomogenization()
 
-    P, start = Problems.problem_startsolutions(Input.StartTarget(
+    P, _ = Problems.problem_startsolutions(Input.StartTarget(
         [x^2+y^2+z^2, x^4+y^4+z^4], [x^3+z^3,y^3-z^3], [rand(ComplexF64, 3)]), homvar=z)
     @test Problems.homogenization(P) == Problems.Homogenization(3)
 
 
     F = Systems.FPSystem(Utilities.homogenize(equations(cyclic(6))))
-    P, start = Problems.problem_startsolutions(Input.TotalDegree(F))
+    degrees = Input.check_homogenous_degrees(F)
+    P, startvals = Problems.problem_startsolutions(Input.TotalDegree(F, degrees))
     @test Problems.homogenization(P) == Problems.NullHomogenization()
-    @test length(start) == 720
+    @test length(startvals) == 720
 
     F = Systems.FPSystem(Utilities.homogenize(equations(cyclic(6))))
-    P, start = Problems.problem_startsolutions(Input.TotalDegree(F), homvar=5)
+    degrees = Input.check_homogenous_degrees(F)
+    P, startvals = Problems.problem_startsolutions(Input.TotalDegree(F, degrees), homvar=5)
     @test Problems.homogenization(P) == Problems.Homogenization(5)
-    @test length(start) == 720
+    @test length(startvals) == 720
 end
